requirements.txt
Django>=3.0,<4.0
django-environ>=0.8,<1.0
# psycopg2>=2.8
gunicorn==20.1.0
psycopg2-binary==2.9.3
dj-database-url==0.5.0
Pillow==8.3.2
selenium==4.22.0
djangorestframework>=3.12,<3.13
seleniumbase==4.28.0
requests==2.31.0
webdriver_manager==4.0.1
flake8>=3.9